Update setup.py to remove version duplication
This change addes my well vetted ast parsing solution to finding the version string that lives in the package source code, without needing to have the package installed, allowing for it to be used in setup.py.

Additionally, I added a minimum python version to the setup call. f-strings are used throughout the code base and therefore we need to be using at least python3.6