Simplify file/path polymorphism
This PR simplifies the handing of files vs path strings (or pathlib objects)
using my file_or_name library. There are several edge cases (related to file
closing, non-positional parameters, and generators) that my library tests.
Therefore I opted to make it as a dependency instead of vendoring it into
the code base. We can revisit this if people want.

Note: The `read_gitattributes` function assumes that the parameter is
only a string and additionally handles what to do when the file is
missing. There for it was not a good candidate to be decorated with this
file opener.