Fix bug in adding tracking to .gitattributes (#66)
* Fix bug in adding tracking to .gitattributes

The git cml track command has a bug where `filter=cml` wasn't added to
lines that matched the provided file path. The issue was that python
strings are immutable, so the `+=` that was used to append `filter=cml`
results in just updating the reference to the iteration variable, the
change isn't persisted.

Additionally, it fixes a bug where adding `filter=cml` (if it had worked) would have been added after the newline.

It also updates the pattern matching comparing the gitattributes pattern
to the model path to use `fnmatch.fnmatchcase`, which evaluates glob
matches, instead of a direct comparison. This means we no longer add
the model path if it is already covered by some other pattern.

* fix typo