Fix location of the model dir in smudge
Currently the model_dir used during smudge is the relative path to the checkout from the repo root, not the converted on that lives in the .git_cml directory. This update changes it to the correct dir, this enables things like `git reset --hard` to update changes to a model and return to the last checked-in state, currently this results in a `FileNotFoundError: [Errno 2] No such file or directory: .../git-theta-test/my-model.pt` error as it tries to parse the directory structure from the actual model file.