WIP: A sketch of plugins for checkpoint support.
This is a sketch of what a plug-in system would look like for
user-defined checkpoint handlers.

It seems pretty low friction, especially because internally we only need
to add new handler to the entry_points dict in setup.py, we don't need
to create an installable package for each checkpoint type.