Adds a TF Checkpoint plugin.
This PR adds a Tensorflow checkpoint plugin based on the checkpoint
format output from `tensorflow.keras.Model.save_weights`. This format
does not include the computational graph.

This PR includes tests for git-theta to read and write a tf checkpoint
and ensure that the values are correct (one test leaves all values
unchanged and another makes modifications).

It also leaves a skeleton for the tensorflow SavedModel checkpoint
format but that is not implemented.

It also updates some of the installation so that not all frameworks are
needed at once and it ensure that tests are only run of the required
framework is installed.