Add support for multiple update types via plugins
This PR adds support for parameter updates, where the parameter value at
timestep t is computed as an update to timestep t-1. The update value is
what is stored on disk instead of the actual parameter at timestep t.
When the parameter at timestep t in reconstituted in a git smudge, it is
computed by applying the update (loaded from disk) to the parameter at
timestep t-1.

We track updates on the filesystem using a slightly updated structure.
Each network parameter still has a `.../params/` directory, but instead
of the parameters living directly in that directory, they live in a
`.../params/updates/${hash}/` directory. This hash is the hash of the
parameters at timestep t, not the hash of the update on disk.

Additionally there is now a `.../params/metadata` file that tracks what
the last update to be made to these parameters was. There is also a
`.../params/updates/${hash}/metadata` file. This tracks what kind of
update was used and what the previous update was.

Now reconstituting via a smudge is done by creating the update plugin
used for the most recent update and calling `.apply` on it. This will get the
value of the parameter after that update was applied. True updates, like
sparse, will recursively get the parameter value at timestep t-1 by
creating and applying the update plugin used in the previous update. It
will then apply it's own update to the result.

Updates that touch all the parameters, for example a dense update, act
as base cases. They don't track what the last update was or call plugin used for the last update, they just
return the parameter values as they were at some timestep i (read
directly from disk).

I opted to implement this recursive update (where each update keeps a
backpointer to the update before it) instead of a list of updates
implementation we has discussed as it allow us to skip the creation of
parameter t-1 when the update at t is not a true update (dense).

The on-disk format for parameters is more complex/verbose now, but the
shared format for different update types makes manipulation easy,
additionally, naming updates by the hash of the resulting parameter
(instead of the update values) should allow for an easy implementation
for things like `git theta compact` which could collapse a series of
sparse updates into a single update.

This PR also adds some tests to the cascading override behavior we
use when determining what update type to use based on various methods of
capturing user input.