Adds a `.name` property to checkpoint handlers.
This PR adds a `.name` property to checkpoint handlers. This name can be
used to look up the checkpoint handler as a plugin. This property lets
us know what handler is being used (for example in logging) even when
the handler selection was done with cascading override (for example the
environment variable was used instead of cli based user input). i.e.
Adding this lets us know that a pickled dict checkpoint handler was used,
even if the actual user input was `None`.