Update the parameter functions to remove torch dep
This PR updates the functions in the `git_theta.params` module to remove
the dependency on pytorch. Instead it operates using numpy arrays.

This also adds tests that we are extracting the right values from
the parameters.

This also updates some of the docstrings on the checkpoint plugin
classes to highlight that the processed model dict should be using numpy
arrays instead of framework native tensors.

When we eventually work with non-standard dtypes (`jnp.bfloat16` for
example) we may need to revisit the initial call to `np.asarray`.