Refactor Update plugins
This PR updates the Update plugin base classes. There is still the
original base-class with overrideable read, write, and apply, but there
is also a new TrueUpdate class which has definitions for read, write,
and apply that work under the assumption that this update is handled
using the previous value. This class includes a few smaller method about
calculating and applying updates to increase code reuse between true
update plugins.

The new implementation allow for writing new checkpoint types really
easily and makes understanding the implementation of the true updates
easier. I looked at trying to include this in the original Base class
but some of the Dense updates implementations of the helper methods
would be strange NoOps and not really make sense.

This PR also updates the format of the parameter level metadata file (it
is now a dict) to allow unification of "get most recent update" when
talking about the whole parameter or the most recent update before the
current update.