Move to Async Read/Writes (#121)
This PR updates the our clean and smudge filters to use async to do I/O
bound operations (mostly tensorstore serialization and running git-lfs-(clean|smudge)). This is done with an async wrapper of subprocesses that mimics subprocess.run. It also includes tools for running an async function on all the k, v pairs in a dictionary so that they will happen concurrently.

The main async entry point is in the clean and smudge filters where the
cleaning (or smuding) function is applied to each leaf in the parameter
tree concurrently.

This also updates the Update API a little bit. This update makes async
possible (i.e. some methods are converted from `def` to `async def`) and
it also makes the API easier for writing new update plugins. Additionally
the serializer is now passed to the update class and used internally
instead of being used externally. Also things like communication with
git-lfs is now done in the write method explicitly instead of being part
of the LfsMeatadata creation.