Add a Low-Rank Update type. (#127)
This PR adds an implementation of Low Rank Updates. In this setting, the
parts of the low rank update are inferred from the difference between
the previous and current parameter using SVD.

This includes tests that verify that rank inference based on the
singular values and also that the application of update results in a
numpy allclose result.