Hotfix for Incremental Updates (#139)
In the original multi-pointer implementation, the writing of the parameter was done as a multi-step process in the clean filter. When I folded that functionality into the IncrementalUpdate class I made a mistake with the semantics of the `param_metadata` parameter. During a clean filter, `param` is the parameter value in the checkpoint right **now**, while `param_metadata` is the metadata from the model that **currently checked in**. This means that when we "get the previous value" we actually want the value that this `param_metadata` represents, as it is previous to the value that is in the checkpoint right now.

Contrast this with the `apply` method. In this case, there is no `param`
vaule, instead there is just `param_metadata` and we want to get the
value of the parameter for that metadata. This means that for an
Incremental Update apply does need to fetch the previous values.

In trying to share as much code as I could, I ended up folding the get
previous metadata functionality into a get previous value method. This
results in write doing a double look up, skipping the true previous
value.

This PR fixes this change by moving the previous metadata fetching out
of the value fetching and only calling it during apply. It also renames
the parameter of the `write` method to `prev_metadata` so it is clearer
going forward.