Add `GIT_THETA_MAX_CONCURRENCY` that can limit the number of async functions running at one time.
When working with larger models (T0 3B, etc.) one can occasionally hit a
"Too many open files" error as async calls to git-lfs, tensorstore and
the like grab file descriptors.

One way to fix this is to change the limit on the number of open files a
process can have (`ulimit -n \d+` on linux) but it might be different on
different systems.

This offers a solution that might be easier for users and would allow us to
ship with a sensible default so users don't install and get hit with a
weird error on the first model they try to checkpoint.

But for now the default is full concurrency.