Add abstract Class Attributes
The PR changes the `name` attribute on plug-ins from being properties to
being class attributes. This makes it possible to access information
about the plug-in (name etc) without having to instantiate it.

It also adds a new class decorator that allows us to enforce the
defining of an attribute in subclasses (i.e. user plugins), similar to
the way we can for object-level properties.