Move scripts to entry points.
This PR updates the binaries from git-theta to live in
`git_theta/scripts/...` instead of `bin/...` and uses the
`console_scripts` entrypoint instead of the `scripts` argument.

The former is the recommended method and is supposed to smooth over some
possible windows issues. Additionally this makes it easier to make
changes to the runnable files in an editable install (you no longer need
to re-run pip install after a change).

There is no user-facing change as the commands `git-theta`,
`git-theta-filter`, etc. still work (they are the entry point names).