Include some commit information in the merge ui
This PR adds some information in the merge UI, it uses the
`GITHEAD_${SHA}` env variable to tell what the other branch is and the
git repo object for information about the current commit.

This summary is shown at the beginning of merging and can be shown again
with the new "context" merge plugin.

This also adds the ability to have "no-op" merge plugins that do not
advance the current parameter. Instead they do something like print the
context and then re-display the menu of actions.