Manual Merge Escape Hatch
This PR allows for a "manual" merge where the merge tools dumps three
copies of the model to disk, the model as it lives on the current
branch, the model as it is on the branch we are merging in, and the
model as it was at the ancestor commit.

The 3 checkpoints can be manually combined, saved to the original path,
and then commited to continue the merge.

This is currently triggered with the `GIT_THETA_MANUAL_MERGE=true` env
variable.