Feat/win tempfile (#225)
* ups

* Add a utility for temp files in tests

The PR adds a wrapper around Python's named temporary file
functionality. When used on windows, there are often permission errors
when trying to open the file by name. As a work around we often open the
tempfile so that it is not deleted after leaving the scope, calling
`.flush()` and `.close()` on it, and then
manually deleting it. This encapsulates that workflow.

This is something we that generally only using during testing, however,
but including it in the main library is the easiest way to import it
in tests. I can figure out a way to move it to the test area if we hate
that.

That PR also updates a few tests to use this new function.

* setup importable test helpers