Update how `git-theta track ${path}` works. (#227)
* Update how `git-theta track ${path}` works.

Currently, when a pattern in `.gitattributes` matches the path in
`git-theta track`, the theta attributes (filter, merge, diff) are added
to that line. This can cause issues as it may result in tracking more
files with git-theta than expected.

However, just adding a new line that is an exact match for the path to
the end of the gitattribute file is not correct either. The *last*
attribute line in the file is the one used by Git. This could result in
`git-theta track` removing an some other attribute that is set for that
file.

This PR updates the way that git attributes are set. If there are
already git attributes set to non-theta values for the file that are used by git-theta
(filter, merge, diff) an error is raised. If these attributes are all
set to git-theta, no new entry is added, even when the entry is a
pattern match instead of an exact match. If the new file has no
attributes set, or attributes that don't overlap with git-theta, then a
new entry is added. Non-overlapping attributes are copied down if they
were set before.

When a line has an attribute set multiple times, the *last* one is used
so we can add the theta filters at the end to override any previous
ones.

Added a `is_theta_tracked` function similar to the one from
https://github.com/r-three/git-theta/pull/214 where the test for if a
file is tracked by Git-Theta is abstracted into a function.

However it is implemented differently as it now handles the subtleties of
what attribute line is active at a given time.

Only the final line that a path matches is used to set attributes and
only the last entry for some key is used. This is now respected.

Git attributes support the "unsetting" of attributes (e.g. "-diff"),
these attributes are correctly copied around, but they aren't currently logically
checked to see if they intersect with git-theta attributes.

* bump setup python version