Git Theta Low Memory Mode
The original file, piped into the filter from git, it stored in a
temporary file before being read by the checkpoint plugin.

When cleaning parameter groups, we free the memory for the group after
it has been written to disk and converted to metadata.

This is the most everything we can do until the dl native formats support
streaming.