Check the error code for `git lfs smudge` (#252)
Currently our git lfs smudge runner never checked the error code. This
lead to a weird error where when git lfs smudge failed, the stdout was
set to the stdin. Thus the lfs pointer file contents was passed along
until it caused an error in the msgpack decoding. This PR updates out
git lfs runner to check the return code and raise an error if it is
non-zero. This makes the error message much more understandable.