Implement the ability to infer the checkpoint type
Currently this is just based on the file path. In order to make the
sniffing extensible, it is implemented as a plugin. In order to do that,
and to avoid having to import a deep learning framework to do sniffing,
the sniffers are separated out into their own files. This lead to the
question of why are checkpoint plugs stored in the main repo? If we move
them to the actual plugins area we can continue to support them while
not needing to install them if a framework isn't going to be used.

Now when you want to git theta with a specific framework you can either
run `pip install git-theta[framework]` or run `pip install
git-theta-checkpoints-framework` for your framework.