Update matrix of options to support macos again.
The new versions of MacOS Runners are ARM based and the setup-python
action doesn't support older python versions (before 3.11) on ARM.

This PR updates the matrix to include both macos-latest (arm) and
macos-13 (the last x86_64 version). It then use excludes to only run new
python versions on ARM and older versions of x86.

We miss out on coverage of someone compiling their own version of < 3.11
python on apple silicon, but I don't think that's too much of an issue.

If we ever need to bump the minimum python version beyond 3.11 we can
remove this.

It also removes the parallelism limiters on the matrix computation (not
sure why I added that, GitHub lets you have infinite parallel runners
for now lol).